services/recurring/requirements.txt
# Recurring Task Service Dependencies (Phase V)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
dapr>=1.14.0
dapr-ext-fastapi>=1.14.0
python-dateutil>=2.9.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
